fix(content): widen Color::lerp/scale channel math to avoid i32 overflow
Color::lerp(a,b,num,den) and Color::scale(num,den) take caller-controlled
i32 ratios. The per-channel math computed (cb-ca)*num (lerp) and c*num
(scale) in i32, so a large num (brighten by a big factor) overflowed:
255 * i32::MAX panics in debug. The den==0 guards were already present;
the overflow was in the multiply.

Switched the channel arithmetic to i64 before clamp(0,255) — overflow-safe
and identical for normal ratios. Surfaced by the robustness lens; added
color_ops_are_total_at_extreme_ratios. 10th and final module in the
raw-i32 overflow class.

PINNED hashes + golden unchanged; 39 content unit tests pass; lib 2358;
robustness 17; clippy --all-targets clean.

+++ izanagi_kit/src/content.rs
@@ -92,7 +92,9 @@ impl Color {
             return a;
         }
         let ch = |ca: u8, cb: u8| -> u8 {
-            let v = ca as i32 + (cb as i32 - ca as i32) * num / den;
+            // i64 so a large `num` cannot overflow `(cb - ca) * num` (channels
+            // are bounded but `num` is caller-controlled).
+            let v = ca as i64 + (cb as i64 - ca as i64) * num as i64 / den as i64;
             v.clamp(0, 255) as u8
         };
         Color {
@@ -137,7 +139,8 @@ impl Color {
         if den == 0 {
             return self;
         }
-        let ch = |c: u8| -> u8 { (c as i32 * num / den).clamp(0, 255) as u8 };
+        // i64 so a large `num` cannot overflow `c * num` (the brighten path).
+        let ch = |c: u8| -> u8 { (c as i64 * num as i64 / den as i64).clamp(0, 255) as u8 };
         Color {
             r: ch(self.r),
             g: ch(self.g),
